refactor CDN deprovision tests to use shared helpers
